* 'dev' of https://github.com/sparta-final/final-project:
  Feat : 피드 수정/삭제 완료
  Style : 단순 변경 #57
  Feat : 헬스장 상세페이지 #57
  Feat : 피드 댓글 작성/조회 연결 완료
  Design : 헬스장 상세 페이지 #57
  Feat : 헬스장 리스트 카카오 맵에 표시 #57
  Feat : 전체 피드 가져오기 service 로직 수정 & 디자인
  Fix : 로그인,회원가입 완료 시 location.replace #58
  Design,Feat : 헬스장 리스트 #57

+++ src/domain/feed/feed.service.ts
@@ -47,12 +47,22 @@ export class FeedService {
 
   /**
    * @description 모든 피드 조회
-   * @author 정호준
+   * @author 정호준, 한정훈
    */
   async getAllFeed() {
-    const allFeed = await this.feedsRepository.find({
-      relations: ['feedsImgs'],
-    });
+    const allFeed = await this.feedsRepository
+      .createQueryBuilder('feeds')
+      .leftJoinAndSelect('feeds.user', 'user')
+      .leftJoinAndSelect('feeds.feedsImgs', 'feedsImgs')
+      // .select(['user.nickname'])
+      .select(['feeds', 'feedsImgs.image', 'user.nickname', 'user.profileImage'])
+      .orderBy({ 'feeds.id': 'DESC' })
+      .getMany();
+    // .getRawMany();
+    // const allFeed = await this.feedsRepository.find({
+    //   relations: ['feedsImgs', 'user'],
+    //   order: { id: 'DESC' },
+    // });
     return allFeed;
   }
 
@@ -68,38 +78,56 @@ export class FeedService {
     });
     return myFeed;
   }
+
+  /**
+   * @description 내 피드 데이터 가져오기
+   * @argument id
+   * @author 한정훈
+   */
+  async updateGetFeed(id) {
+    return await this.feedsRepository.findOne({
+      where: { id: id },
+      relations: ['feedsImgs'],
+    });
+  }
+
   /**
    * @description 내 피드 수정
-   * @param feedId @param file @param updatefeedDto @param user
+   * @param feedId
+   * @param updatefeedDto
    * @author 정호준
    */
-  async updateFeed({ feedId, file, updatefeedDto, user }) {
-    await this.checkUser(feedId, user);
-    const existFeed = await this.feedsRepository.findOne({
-      where: { id: feedId },
+  async updateFeed({ id, updatefeedDto, user }) {
+    await this.checkUser(id, user);
+    await this.feedsRepository.update(id, {
+      content: updatefeedDto.content,
     });
-    const findFeedImg = await this.feedsImgRepository.findOne({
-      where: { feedId: feedId },
-    });
-    const queryRunner = this.dataSource.createQueryRunner();
+    // 이미지 함께 수정 가능하게 하려면 위에 코드 대신 아래코드 작성
+    // const existFeed = await this.feedsRepository.findOne({
+    //   where: { id: feedId },
+    // });
+    // const findFeedImg = await this.feedsImgRepository.findOne({
+    //   where: { feedId: feedId },
+    // });
+    // const queryRunner = this.dataSource.createQueryRunner();
 
-    await queryRunner.connect();
-    await queryRunner.startTransaction();
+    // await queryRunner.connect();
+    // await queryRunner.startTransaction();
 
-    try {
-      await this.feedsRepository.update(feedId, {
-        content: updatefeedDto.content ? updatefeedDto.content : existFeed.content,
-      });
-      await this.feedsImgRepository.update(findFeedImg.id, {
-        image: file ? file.location : findFeedImg.image,
-      });
-      await queryRunner.commitTransaction();
-    } catch (err) {
-      await queryRunner.rollbackTransaction();
-      throw err;
-    } finally {
-      await queryRunner.release();
-    }
+    // try {
+    //   await this.feedsRepository.update(feedId, {
+    //     content: updatefeedDto.content ? updatefeedDto.content : existFeed.content,
+    //   });
+    //   await this.feedsImgRepository.update(findFeedImg.id, {
+    //     image: file ? file.location : findFeedImg.image,
+    //   });
+    //   await queryRunner.commitTransaction();
+    // } catch (err) {
+    //   await queryRunner.rollbackTransaction();
+    //   throw err;
+    // } finally {
+    //   await queryRunner.release();
+    // }
   }
 
   /**
@@ -107,14 +135,16 @@ export class FeedService {
    * @param feedId @param user
    * @author 정호준
    */
-  async deleteFeed({ feedId, user }) {
-    await this.checkUser(feedId, user);
+  async deleteFeed({ id, user }) {
+    await this.checkUser(id, user);
     const findFeedImg = await this.feedsImgRepository.findOne({
-      where: { feedId: feedId },
+      where: { feedId: id },
     });
+    console.log('✨✨✨', 'findFeedImg', findFeedImg, '✨✨✨');
     const findComment = await this.commentsRepository.findOne({
-      where: { feedId: feedId },
+      where: { feedId: id },
     });
+    console.log('✨✨✨', 'findComment', findComment, '✨✨✨');
 
     const queryRunner = this.dataSource.createQueryRunner();
 
@@ -124,8 +154,10 @@ export class FeedService {
     //TODO: 댓글삭제 시 한개만 삭제가 아닌 feedId를 가진 전체 삭제 구현해야 함
     try {
       await this.feedsImgRepository.delete(findFeedImg.id);
-      await this.commentsRepository.delete(findComment.id);
-      await this.feedsRepository.delete(feedId);
+      if (findComment) {
+        await this.commentsRepository.delete(findComment.id);
+      }
+      await this.feedsRepository.delete(id);
       await queryRunner.commitTransaction();
     } catch (err) {
       await queryRunner.rollbackTransaction();
@@ -163,15 +195,35 @@ export class FeedService {
     });
     return createComment;
   }
+
   /**
-   * @description 피드 댓글 조회
+   * @description 피드 작성자 조회
    * @param feedId
    * @author 정호준
    */
+  async getCommentUser(feedId) {
+    return await this.feedsRepository
+      .createQueryBuilder('feeds')
+      .leftJoinAndSelect('feeds.user', 'user')
+      .where('feeds.id = :feedId', { feedId })
+      .select(['feeds', 'user.nickname', 'user.profileImage'])
+      .getRawMany();
+    console.log(feedId);
+  }
+
+  /**
+   * @description 피드 댓글 조회
+   * @param feedId
+   * @author  한정훈
+   */
   async getAllComment(feedId) {
-    return await this.commentsRepository.find({
-      where: { feedId: feedId },
-    });
+    return await this.commentsRepository
+      .createQueryBuilder('comments')
+      .leftJoinAndSelect('comments.user', 'user')
+      .leftJoinAndSelect('comments.feed', 'feed')
+      .where('comments.feedId = :feedId', { feedId })
+      .select(['comments', 'user.nickname', 'user.profileImage', 'feed.content', 'feed.userId'])
+      .getRawMany();
   }
   /**
    * @description 피드 댓글 수정
